updated score message to allow players to play online together

+++ Assets/Scripts/UI/ScoreMessage.cs
@@ -35,8 +35,8 @@ namespace nickmaltbie.IntoTheRoots.UI
 
             if (enabled)
             {
-                int winner = loop.Winner;
-                if ((long) winner == (long) NetworkManager.Singleton.LocalClientId)
+                long winner = loop.Winner;
+                if (winner == (long) NetworkManager.Singleton.LocalClientId)
                 {
                     winMessage.text = $"You Won!";   
                 }
